[BatchRenameJob] Use KJob::Items when reporting the progress info
A batch rename job is usually fast (direct renaming), calculating the
number of individual files and dirs would mean stat'ing which would make
things slower for little gain. Instead, per dfaure's suggestion in the bug
report, use "items" as a neutral term to represent files and dirs, this
way the notification message would be "correct" when reporting "Job Finished, x
of y items".

Also note that if the user is using Dolphin, or other KIO based file
managers, that data is already visible in the status bar when the user
selects the files/dirs to be renamed.

BUG: 422098