[Trash] Remove trashinfo files referencing files/dirs that don't exist
If the user moves a file to trash, the deletes that file without using
kio_trash (e.g. with rm), then the relevant trashinfo file should be
deleted the next time we try to list/stat that trashed file; otherwise
kio_trash keep trying to stat that file every time the Trash dir is
stat'ed/listed.

The testtrash unit test still passes.

Thanks to meven for pointing out how to fix this bug.

BUG: 422189
FIXED-IN: 5.72