[kcookiejar] Fix reading "Accept For Session" cookie setting
Since commit 23874cab, the "Cookies" KCM writes the option as "Accept For Session" (with spaces), probably by accident, which broke reading the saved setting.
The KCM has been adjusted later on (in commit efeda6c1) to accept both spellings when loading the settings.

But the daemon itself still only understands "AcceptForSession" (without spaces), so it still asks the user what to do with cookies in this case.

This patch applies the same fix to the daemon, so that changing the setting in the KCM actually works as expected.

CCBUG: 386325