kfilewidgettest: prevent the unit test from altering ~/.config/kdeglobals
When running kfilewidgettest, the saved view settings of the open-file
dialog not loaded from kdeglobals, therefore it uses the default view
settings, then when the dialog is "accpeted" during the unit test, those default
setings are written to ~/.config/kdeglobals, which overrides the user
settings.

To test (before applying this patch), set the view mode to "compact" in the
open file dialog, run this unittest then check the dialog again, it'll be
using the "details" view mode.