[KUrlCompletion] Don't append / to completed folders
This is a quick fix to a usability issue.

Autosuggestions are made after / is appended to the folder name and pressing enter completes the autosuggetion. These 2 things are not wrong on their own, but they mess with people's heads when combined in this way. This change prevents the autosuggetion from being started when completing directories since it's only activated after the / is appended.

A better, but not so quick fix would be to make it so pressing enter only completes an autosuggestion after a character has been added after the /