Introduce OpenOrExecuteFileInterface for handling opening executables.
When opening various types of "executables" (native binaries, .exe files, shell scripts or .desktop files), show a dialog to ask the user whether to run/execute it or open it as text (in the case of text-based ones). This uses ExecutableFileOpenDialog for KIOWidgets.

This is modelled after OpenWithHandlerInterface.

Add unit test to openurljobtest.

Add unit test to openurljobtest.