KFilePlacesModel: add default places when upgrading from older version
The unittest is complicated by the interaction between KFilePlacesModel
instances via the KBookmarkManager singleton, and the fact that KBookmarkManager
uses KDirWatch internally to react to changes in the places xbel file
(often after a small delay). Maybe a separate unittest would have been simpler;
or we could try to completely delete the KBookmarkManager instance.