smb: rejigger guest allowed check and publish as areGuestsAllowed
this was previously only used inside setGuestPermissions to check if
setting it to Allowed was even ... allowed.
in addition to that behavior we'll now provide it as a global bool
getter on the singleton. this allows UIs to check ahead of time if guest
access is allowed and present the state accordingly