Revert "Avoid systemd launched applications from closing forked children"
Whilst this commit was useful in finding issues, and the long-term right
direction, it encounters some issues:

* There are edge cases where the envionment is not synced to systemd,
leading to nothing launching. This is resolved in new Plasma, but we
have an issue with Plasma 5.19 and new frameworks.

* KSMServer had a bug where it's env wasn't synced to the activation
env. This is resolved in new Plasma, but we have an issue with Plasma
5.19 and new frameworks.

* An upstream issue means cgroups leak in some edge cases.

* There is a bug with KProcessRunner and startup ID getting into the
environment correctly.

Ultimately the move to services from scopes may be technically better,
but we have teething bugs that need time to resolve. It doesn't bring
tangible benefits we need in the near future.

I would prefer we revert for now to play safe. We don't want any
regressions.

This reverts commit a9db1110bd306142dd12e35c6025608f73f25dbe.
This reverts commit a7bf244e5ff65c7fad20812dda99e84590c47045.
This reverts commit c5ae27d119b3bfac2f5ab7a35540f0fe260ee2b3.
This reverts commit 57438d6dfe4ecb1ae022775401f5c7bcbf2a93da