Change name of environment variable used for new systemd launch feature
The use of cgroups was guarded by an environment variable for our first
rollout to perform it safely with no regressions.

A massive overhaul landed which exposed some existing bugs in Plasma and
the surrounding stack. This patch renames the environment so we can do
the slow-roll out again ensuring user experience stays regression free.