put minimumkeepsize actually in the netpref KCM
I didn't even know it was a thing! by the looks of it the setting only
applies to retention of .part files on aborting transfers (that is,
files must be at least this size so they get retained on aborting
transfers).
so the setting is in a form inside a widget and that widget's enabled
state is contient on partial marking being enabled

CCBUG: 419987