KDirOperator: don't call setCurrentItem on an empty url
A previous commit re-enabled dir highlighting, which is supposed to
highlight the dir we've just left when going back/up. However when a
KFileWidget is first created/opened there is no such item to highlight.
If setCurrentItem is called with on an empty url, KDirModel::expandToUrl
would call expandAllParentsUntil on that url which would show this warning:
'protocol mismatch: "" vs "file"'.

This also fixes an issue where if you open a file in kate, then invoke the
open file dialog the currently open file isn't selected/highlighted in the
view.

BUG: 425163
FIXED-IN: 5.75