KUrlNavigator: always use "desktop:/" not "desktop:"
Navigating to "desktop:" in e.g. dolphin, then creating a new folder would
lead to creating "/home/user/DesktopNew Folder", that's because the "path"
segment in the url is empty. The fix is in two parts, one here by ensuring
we always go to "desktop:/", even if the user typed only "desktop:". Do
this for all KProtocols that have class ":local"; something like
"file:somepath" would treat somepath as a relative path (in dolphin,
relative to the current dir), we lose that "functionality", but I think
it's too obscure, so losing it is not a big deal.

The other part is in kio_desktop in plasma-workspace.

Add unit test.