kio_trash: fix the logic when no size limit is set
If a size limit is set, we use that to calculate the percent of the total
partition size that trash can use; however if the size limit isn't set,
then the available free space on the partition is the limit.

Thanks to Patrick Silva for figuring it out in the bug report.

BUG: 426704
FIXED-IN: 5.76