KFilePlacesModel: ignore hidden places when computing closestItem
KUrlNavigator uses closestItem to get the closest bookmark so as to show it
as the first button in the url bar.

For example, open Dolphin on the Documents bookmark, "Documents" is shown
as the firstbutton in the urlbar (breadcrumb); now hide that bookmark, the
urlbal should now show "Home > Documents".

BUG: 426690
FIXED-IN: 5.76