KFilePlacesItem: hide kdeconnect sshfs mounts
https://phabricator.kde.org/D23694 adds SSHFS support to Solid, which
makes all SSHFS mounts show up as network devices in the places model.

KDE Connect uses SSHFS to browse the phone's files but it already
provides a dummy kdeconnect:// kioslave to access the sshfs mount from
the places panel.

This patch hides the SSHFS mount from kdeconnect because, as explained,
it is not needed and it would also show the internal mount point
(/run/user/...) as the place visible name.

We still give users the ability to explicitly show this place,
should they wish so.