Handle KCMs in OpenUrlJob with KService API
Addition to https://invent.kde.org/frameworks/kservice/-/merge_requests/16.
The Exec lines are implicitly defined to avoid redundant values and
workarounds in different places. We are actually using the KService
API inside of this logic and this commit just reads the exec
value from KService instead of the raw desktop file.