OpenUrlJob: minor code refactoring
As per dfaure's suggestion, split the common code out of handleBinaries to
a helper method. Note that since m_runExecutables is modified by the
showOpenOrExecuteFileDialog method, and that for "binaries" the dialogue
will only show Execute/Cancel buttons (but not "Open") we only call
handleBinariesHelper if m_runExecutables is "true", so the "For security
reasons" error message won't be shown, which is the expected behaviour
since we already asked the user with the openorexecute dialog.