[PreviewJob] Allocate sufficient memory for SHM segment
cacheWidth/Height is clamped to 256x256 when some of the the thumbnails
are saved, while the unsaved thumbnails are requested with the original
(larger) size.

Dolphin and the filewidget disable caching for the large previews, and
the iconviews are limited to 256x256, so typically this does not happen,
but when the PreviewJob is run without disabling caching using
`setScaleType`, the thumbnailer slave will write out-of-bounds.

CCBUG: 427865