expose user context help field in kpasswdserver
this adds a new extra field which may be set to expose contextual help
on the username field of KPasswordDialog.
this allows auth consumers such as slaves to give the user more
contextual information on what they expect the username to be. for
example windows logon names and by extension smb user names support a
range of different input formats that the user may need to use depending
on the authentication environment.