CopyJob: add unittests for skipping during copying
To my surprise, this wasn't actually tested.
Confirmed by Jenkins' code coverage view for copyjob.cpp (great
feature! That's a TODO list in itself...).

This detected a bug in the total amount of files, I'll fix it
in a later commit (this way we see better what the fix changes).