Support for xattrs on kio copy/move
Summary:
This patch adds support to KIO to preserve xattr of files and directories on copy or move.
Working for copy/move including overwrite and rename cases.
Not working for "write into" when coping folders have name conflict. I'm against adding to this user case as the name of the action don't imply overwrite.

Tested only on Linux.

FEATURE: 116617

Reviewers: dfaure, chinmoyr, bruns, #frameworks, tmarshall, usta, cochise

Reviewed By: dfaure, usta

Subscribers: usta, scheirle, tmarshall, arrowd, cfeck, bruns, phidrho, dhaumann, funkybomber, abika, pino, davidedmundson, ngraham, atha.kane, spoorun, nicolasfella, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17816