CopyJob: Include skipped files in progress calculation when renaming.
We want "1/2 files moved" but 100% progress. In the normal case this is
handled by the "bytes processed" calculation, but for direct-renaming
this is based on files. So we don't want setProgressUnit(Files) after
all, we want our own percent calculation so that we can get to 100%
even when skipping files.

This makes the behaviour more consistent in the two cases: moving
on the same partition or moving between partitions.

Detected by CI because there $HOME and /tmp are actually on the same
partition.