[RFC] Remove SMB KCM
The KCM allows only to configure default username/password for SMB shares.

The only user of that information is the SMB KIO slave.

I would assume that shares usually have unique username/password combinations and the user expects to enter them on first use.
Therefore I'm not sure the concept of a default username/password makes much sense.

Existing configurations are not affected by this.