Simplify code for RequiredNumberOfUrls
We can make life easier for us if we just read the
value as what it is: A list of int.
Then we can do a simple contains. Values smaller than 1
would be ignored as before, but it does not make any sense
to write them in the service file ;)