KUrlNavigator: resolve relative dir paths
When the current url is e.g. /path/to/foo/, and the user replaces all the
text in the url box with "folder", which happens to be a dir in /path/to/foo/,
upon pressing Enter, we resolve the text to /path/to/foo/folder; this seems
more useful.

Hide the new method, uncommittedUrlAsync(), in the d-pointer; the original
method, ucommittedUrl(), doesn't seem to be used by any KDE code other than
the KUrlNave unit tests, but that can still work if we make it a member of
the Private class and make the unit test a friend... etc; the public
uncommittedUrl() should probably be deprecated.

Add a unit test.

BUG: 319700
FIXED-IN: 5.78