KFileWidget: allow files that begin with a ':' to be selected
The issue was that a ':' at the beginning denotes a Qt Resource and in that
case QDir::isAbsolutePath() would return true. The bug is fixed by guarding
against that condition in slotOk().

Also guard other QDir::isAbsolutePath() usage the same way.

Since this logic will be used in various places use a helper function,
isAbsoluteLocalPath() (thanks to dfaure for coming up with that name),
in pathhelpers_p.h.

Extend the unit tests.

BUG: 322837
FIXED-IN: 5.78