Remove knetattach item from the remote:// ioslave's view
This commit removes the knetattach item from the remote:// ioslave's
view entirely.

knetattach lives in Plasma, so this amounts to an unmarked dependency on
Plasma within KIO, which is not appropriate.

Furthermore, Having a desktop entry in the view of the ioslave that
launches an app is kind of a weird UX, and its functionality is largely
unnecessary bceause you can just connect to the server by its URL and be
prompted for your username and password. If you want to create a permanent
link to the server, you can add it to the places panel or use a shortcut.

However for those who would like to continue using knetattach, I plan
to add a contextually-aware button in Dolphin to open the app, and if
that happens without this, we will have two redundant ways to launch
knetattach from the remote:// view.

BUG: 430211
FIXED-IN: 5.78