Do not assert if KIO::rawErrorDetail() is given a URL with no scheme
If the 'reqUrl' points to a URL that is not null but has no scheme,
e.g. QUrl("foo"), then QUrl takes it as as a valid relative URL.
However, the blank scheme is passed to KProtocolInfo::protocolClass()
which asserts.

This can be triggered by, e.g. "konqueror --part webenginepart foo"
where the "foo" generates a relative URL.

Check that the scheme is not empty before calling protocolClass().

CCBUG:393496