Guard QDir::isRelativePath() from paths starting with ':'
Here we use isAbsoluteLocalPath() (in reverse, i.e. isRelative ==
!absolute); This is sort of an addendum to 30abfc8c9a1cba.

Note that the upstream QDir code implements isAbsolutePath() in terms of
"it's absolute if it's not relative". It has the same issue with paths
starting with ':', since ultimately it uses the QFileSystemEngine backend
which treats paths starting with ':' as absolute paths since it considers
such paths as denoting a Qt Resource (qrc).

As dfaure pointed out in [1], this change isn't really needed in
DesktopExecParser.

[1] https://invent.kde.org/frameworks/kio/-/merge_requests/269