webshorcuts: Port the settings to KConfigXT
It uses the automatically generated `WebShortcutsSettings` class to
load, save and set defaults. Some small refactoring has also been done
to simplify the logic.

Not a thorough port, as it does not affect the UI or other automatic
goodies, but it's a starting clean-up point for further work.

QA though `kiowidgets-kurifiltertest-...-separator` unit tests for
no-regressions