kcm/webshortcuts: Fix Reset button
The WebShortcuts KCM is designed as a loader which can have
multiple modules on it. The implementation propagates the main
events to the submodules (currently it's just `kuriikwsfilter`)

The `load()` method implementation was commented out, since
sub-modules automatically call load() on first show. But subsequent
calls need to be propagated to make the `Reset` button work