RenameDialog: add new flags RenameDialog_Source/DestIsDirectory
Per dfaure's suggestion, introduce a RenameDialog_Options flag that e.g.
Copyjob can set to indicate the source is a directory; this is more
efficient since the CopyJob already has all the necessary info about the
source and dest, and means we don't have to use another blocking or
non-blocking stat call.

To eliminate confusion, the new flags are RenameDialog_DestIsDirectory and
RenameDialog_SourceIsDirectory; and the old RenameDialog_IsDirectory is now
deprecated.