KProcessRunner: only emit processStarted() signal once
As discussed with dfaure on IRC, processStarted() signal should only be
emitted once from setPid(). In this code "!m_pid" is used as "this hasn't
been run before" condition.

Since SystemdProcessRunner::handleProperties() gets called more than once,
use a similar check so that it only calls KProcessRunner::setPid() once.

This fixes a crash in applicationlauncherjob_servicetest, which
incidentally passes on the CI and when using kdesrc-build; for the crash
to happen systemd needs to be actually running/used by KProcessRunner.

This reverts commit aa4e0645bb4df111db8c8cef54da3ef18b94ca60.