Compile benchmark tests, but don't run them automatically with ctest
As discussed in [1], benchmark unit tests are useful when benchmarking
performance, but not if no one is looking at the results (either when run
locally or on the CI); change such tests to be always compiled, so as not
to bit-rot, but don't run them automatically when using e.g. `ctest` or
`make test`.

This was first propsed by Meven in https://phabricator.kde.org/D26468

[1] https://invent.kde.org/frameworks/kio/-/merge_requests/301