[KDynamicJobTracker] Support KUiServerV2JobTracker
Track the KUIServer DBus service and check (unfortunately through DBus
introspection) whether it supports API version 2 and cache it until the
service owner changes or disappears. This way only blocking DBus call
is only done initially and subsequent jobs are block-free.

KUiServer is also part of Plasma these days, so there's no need for asking it
about whether a job view is present. This cannot really be changed for V1
but for V2 we can skip this.