Fix lockup when renaming a file from properties dialog
4c12899369f43009269d0fdb80964483365bb922 changed how we show the rename dialog.
Before we would exec() it and now we call show(). This causes a lockup when
renaming a file from the properties dialog to a filename that already exsist.
Optimally we would change KPropertiesDialog to make it async and not use nested
event loops but this fixes the lockup for now.
BUG:431902