Expand test for local files
Some code paths go into DesktopExecParser with the URL in the form
QUrl("/home/david") instead of the correct QUrl("file:///home/david/").

Arguably this is a bug in the calling code, but something seen in the
wild.

KIO doesn't handle this in the test of whether we need kio-fuse is
broken. url.isLocalFile() returns false, and as such we don't see the
program as being able to load this URL.

Testing done:
 - Load a local file from dolphin in firefox. It loads correctly
 - It now appears in "recent" files in the taskbar context menu
- Previously opening that link would open via kio-fuse, mounting
file://
 - now it opens the file normally