ApplicationLauncherJob: improve handling of KServiceAction
As pointed out by davidre on IRC, the ctor that took a KServiceAction
calls KService::setExec() which clears the entryPath() of the KService;
that meant the isAuthorizedDesktopFile() check was by-passed.

To fix the issue, cache the entryPath() of the service before calling
setExec(), which should be OK since it's an Action from the same .desktop
file.

This change also required to change KProcessRunner so as to pass the
cached service entryPath() to its methods.