Fix icon for link to address (Url)
If linking to a local file, use an icon matching the mime type of the Url
the link is pointing to, instead of the folder icon.

Test: Right click on the desktop, Create new -> Link to address (URL),
select a local file. Without this patch always a folder icon is shown,
regardless of the mime type the desktop file link is pointing to. With the
patch applied, the correct icon is shown.

Bug: 428869