don't try to get mimetypes on empty urls
the assert is pre-dating frameworks but really makes no sense. if we
want to assert that items don't have empty urls we should do it in the
setters, but we don't. in fact, we assert this constraint nowhere else,
making me doubt the assertion as a whole.

@vladz has dolphin fall over the assertion every once in a while ->
change it to an early return condition instead. Surely an empty url has
no mimetype.