Add default args to AskUserActionInterface virtual methods
I used default args only for the methods that were overriding (in
WidgetsAskUserActionHandler), but it's useful for the virtual methods to have
default args too, because anything in KIOCore can still call those methods
without having to pass all the args (e.g. in the upcoming port of FileCopyJob
to use the new interface, askUserRename() will be used and FileCopyJob doesn't
pass the various QDateTime args).

This doesn't break SC/BC, quoting dfaure:
"There are no overloads which would become ambiguous, and this doesn't change
the method signature encoded into the symbol."