askuseractioninterface: add TODOs for not using QWidget anymore in KF6
The parameter isn't just for parenting the dialog, it's also used by e.g.
KDirOperator to find out if it is the target for that signal.
So this is a case where we don't want to change to QWindow*, there could
be multiple widgets in the same window, better use a QObject* that
KDirOperator (and others) can set to "this".

Prompted by https://phabricator.kde.org/T11821 (moving FileUndoManager
to kiogui). But of course FileUndoManager::UiInterface is
widget-specific so it would stay in KIOWidgets. So another solution is
to move askuseractioninterface to kiowidgets. But who knows what QML
might need in the future :)