Add KCoreDirLister::jobError signal
Deprecate the KCoreDirLister::handleError() virtual method, in favour
of this new signal.

Deprecate KDirLister::autoErrorHandlingEnabled/setAutoErrorHandlingEnabled
and provide this in KCoreDirLister instead.

The end goal for KF6 is to get rid of the tiny KDirLister subclass of
KCoreDirLister (and rename KCoreDirLister to KDirLister), and then
move KDirModel to kiogui. The only thing left blocking that is
the jobStarted() virtual, to associate a window with new jobs,
this could become a signal too, but how to name it, if not jobStarted()
:-). I also think most users of KDirLister will forget to connect to it,
but at least in KDirModel we could add a setWindow method.

KF6 task: https://phabricator.kde.org/T11821