When changing requestMimeTypeWhileListing to on, clear cached items
If we do not clear cached items, we risk returning cached items that do
not have a mime type, which would mean we end up doing the mimetype
resolution on the client side again, which breaks the request
funcitonality. Since the other way around does not pose a problem, we
only do this when changing from off to on.

This corner case is most likely to be hit in the unit test, since
in normal application use you'd either always or never set this.