FileProtocol: set UDS_SIZE to 0 for /dev, /proc and /sys
These are virtual filesystems, sizes of them and their children aren't useful
for the user. This also fixes a bug where the size of /proc/kcore is
reported as 128TiB in the properties dialog (stat() system function reports
that size, which may be technically correct but not confusing for users).

BUG: 298133