Always compile dataprotocol ioslave to be run "in-process on the client side"
Basically this removes remnants of DATAKIOSLAVE, this was used to control
whether dataprotocol was compiled as a standalone kioslave (kinstance) or
not. Given that kinstance.h has been gone for a long time, this code path
hasn't been compiled in ages (seems no one noticed or missed it), so I
assume it's redundant.