PreviewJob: Create a larger SHM when necessary
It's possible that during the lifetime of a PreviewJob thumbnails with
different sizes are created. The SHM was created to match the size of the first
(non-empty) thumbnail, so any larger ones after that would reuse the SHM with
insufficient size.

BUG: 430862
(cherry picked from commit bc02b2528c97eab0961b3ad608a4914d3ce4b46c)