Remove the Browser Identification KCM
It's purpose is to control the user agent used by kio-http.

Very few things use kio-http these days, not even Konqueror with the
webengine part.

For the remaining use cases (kioclient, webdav kio worker, etc.) this
kind of configurability is overkill.

Moreover being part of systemsettings this gives a false impression of
globality. A user might change something in there and wonder why it does
not apply to Firefox or Falkon.

Given the questionable usefulness and the potential for confusion I
propose to remove it entirely.