Remove the cache KCM
Same reasoning as for
https://invent.kde.org/frameworks/kio/-/merge_requests/438.

It is very targeted towards KIO-based browsers which don't really exist
any more.

The setting is also global to everything that uses KIO and it's a bit
questionable whether for example webdav browsing in Dolphin should
automatically have the same cache settings as network access in Choqok.

It being part of system settings is also confusing for users since it
gives the false impression of true globality. They might expect it to
affect Firefox or Webengine-based Konqueror which it does not.