kio_file: pass the absolute path to QMimeDatabase::mimeTypeForFile()
Otherwise detecting the mime type based on the file content may fail and
return application/octet-stream.

And pass the whole url to createUDSEntry(), less QFile::decodeName/encodeName()
in the middle is better and less error prone.

Note that without this change a MimeTypeFinderJob could end up failing to
find the mime type of a local file based on the file contents.