TCPSlaveBase: deprecate socket() and introduce tcpSocket()
It looks like the socket() method returns a QIODevice* is mostly due to
historical reasons (from the kdelibs days); and it's also obvious that most
sub-classes of TCPSlaveBase that use that method, cast the pointer returned
by it to QSslSocket*, so that they can e.g. connect to the
QAbstractSocket::connected() signal.

So deprecate socket() and add a tcpSocket() method that returns a
QAbstractSocket*; the reason for not returning a QSslSocket (the type of
d->socket is actually QSslSocket) is, as Volker explained to me on IRC, since
socket() returns a QIODevice rather than a KTCPSocket, which meant that the
latter stayed as an implementation detail, porting to QSslSocket was made
easier, so we go one step down the hierarchy, and return a QAbstractSocket*.

Port HTTPProtocol to the new method.