AcessManagerReply: use PMF signal/slot syntax
Looking at how AccessManagerReply is used, the SimpleJob arg in that ctor is
either a StatJob, or a (derived class of) TransferJob, or just a SimpleJob,
(i.e. a job that doesn't have e.g. redirection() signal...etc).

I think this is more "precise", instead of relying on QObject's virtual methods
doing the searching we do it ourselves by casting to the correct KIO::Job class.