New job: KTerminalLauncherJob
This replaces KToolInvocation::invokeTerminal,
and builds on top of (actually, inherits) KCommandLauncherJob.

And this time it comes with unittest and command-line test program.

The Windows code path is especially interesting. It's all ported
from ktoolinvocation_win.cpp. The part in kterminallauncherjob.cpp
was fully tested on Linux (by swapping the ifdefs everywhere).
The Windows-specific flags and handling of arguments in KProcessRunner
needs testing on Windows though.

KF6 task: https://phabricator.kde.org/T14329